Optimized the select query

+++ ServerSharing.Data/SelectResponseData.cs
@@ -11,7 +11,7 @@ namespace ServerSharing.Data
         [JsonProperty("downloads")] public ulong Downloads { get; set; }
         [JsonProperty("likes")] public ulong Likes { get; set; }
         [JsonProperty("rating_count")] public ulong RatingCount { get; set; }
-        [JsonProperty("rating_average")] public double RatingAverage { get; set; }
+        [JsonProperty("rating_average")] public uint RatingAverage { get; set; }
         [JsonProperty("my_rating")] public sbyte? MyRating { get; set; }
         [JsonProperty("my_like")] public bool MyLike { get; set; }
         [JsonProperty("my_download")] public bool MyDownload { get; set; }
